+++ gulpfile.js
@@ -59,6 +59,7 @@ gulp.task('js-merge', function () {
     './src/43rollup.js',    
     './src/44defcols.js',
     './src/45union.js',
+    './src/46apply.js',
   	'./src/50expression.js',
     './src/52linq.js',
     './src/55functions.js',